Fix compression in H5Writer
When using the 'from_file' option in H5Writer, the default behaviour is to copy the compression of the input file. However, if we instead pass 'compression='gzip'' this should overwrite. In the current code, this is not the case as we are instead checking the kwargs for the name of the compression, not the word 'compression' itself. This fixes this.